u-boot-2010.06: add support for more hisilicon chipset.
including hi3516a/hi3516cv200/hi3518ev200/hi3521a/hi3531a/hi3536.

+++ boot/uboot/u-boot-2010.06/include/fat.h
@@ -184,11 +184,11 @@ typedef struct dir_slot {
 typedef struct {
 	__u8	fatbuf[FATBUFSIZE]; /* Current FAT buffer */
 	int	fatsize;	/* Size of FAT in bits */
-	__u16	fatlength;	/* Length of FAT in sectors */
+	__u32	fatlength;	/* Length of FAT in sectors */
 	__u16	fat_sect;	/* Starting sector of the FAT */
-	__u16	rootdir_sect;	/* Start sector of root directory */
+	__u32	rootdir_sect;	/* Start sector of root directory */
 	__u16	clust_size;	/* Size of clusters in sectors */
-	short	data_begin;	/* The sector of the first cluster, can be negative */
+	int	data_begin;/* The sector of the first cluster, can be negative*/
 	int	fatbufnum;	/* Used by get_fatent, init to -1 */
 } fsdata;
